+++ BaragonUI/app/components/common/modalButtons/RemoveUpstreamsModal.jsx
@@ -1,7 +1,8 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import { RemoveUpstreams } from '../../../actions/api/services';
+import { SubmitRequest } from '../../../actions/api/requests';
+import Utils from '../../../utils';
 
 import FormModal from '../modal/FormModal';
 
@@ -49,9 +50,18 @@ class RemoveUpstreamsModal extends Component {
   }
 }
 
+const buildRequest = (ownProps, data) => ({
+  loadBalancerService: ownProps.loadBalancerService,
+  noValidate: data.noValidate,
+  noReload: data.noReload,
+  loadBalancerRequestId: Utils.buildRequestId(ownProps.loadBalancerService.serviceId),
+  addUpstreams: [],
+  removeUpstreams: ownProps.upstreams,
+});
+
 const mapDispatchToProps = (dispatch, ownProps) => ({
-  removeUpstreams: (data) => dispatch(RemoveUpstreams
-    .trigger(ownProps.loadBalancerService, ownProps.upstreams, data.noValidate, data.noReload))
+  removeUpstreams: (data) => dispatch(SubmitRequest
+    .trigger(buildRequest(ownProps, data)))
       .then(response => (ownProps.then && ownProps.then(response)))
 });
